Adjusted msg size to 15400.
Added offset and gain information to the message.

+++ Matlab/averageSamples.m
@@ -8,7 +8,7 @@ outSignal=zeros(1,L/2);
 
 j=1;
 for i=1 : L/2
-   outSignal(i) = uint16(mean(inSignal(j:j+1))); %uint16
+   outSignal(i) = (mean(inSignal(j:j+1))); %uint16
    j = j + 2;
 end
 end
